refactor(zdu): give the upgrade-result stores named classes
Addresses review feedback: the two DataHubUpgradeResultStore
implementations were anonymous classes returned from static of(...)
factories on the interface itself, which is an awkward shape to inject
and hid which transport a caller was getting.

Split them into named classes and let callers name the one they want:

  GMS  -> new EntityServiceUpgradeResultStore(entityService)
  MAE  -> new EntityClientUpgradeResultStore(systemEntityClient)

Dropping the factories also lets the interface stop importing
EntityService and SystemEntityClient. Those imports existed only to
offer the two constructors, which put a datasource-bound type on the
signature of the abstraction that exists precisely so the standalone MAE
consumer does not need a datasource.

No behaviour change: same two implementations, same call sites, same
transport selection via entityClient.impl.

(cherry picked from commit 7728ef08d3a55f79cc9fdd115d47b4f54f6a4fd2)
